Added a int cast to VEC_length.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/cilkplus@183215 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/cilk.c
@@ -1155,7 +1155,7 @@ cilk_remove_annotated_functions (rtx first)
 	    }
 	}
     }
-  for (ii = 0; ii < VEC_length (rtx, rtx_delete_list); ii++)
+  for (ii = 0; ii < (int)VEC_length (rtx, rtx_delete_list); ii++)
     remove_insn (VEC_index (rtx, rtx_delete_list, ii));
   
   return;
